refactor(@leav/ui): extract LinkModal to be used in other contexts
